+++ Source/Origin.cs
@@ -7,38 +7,30 @@ namespace Aksio.IngressMiddleware;
 
 public static class Origin
 {
-    const string OriginHeader = "x-ai-origin-uri";
     const string OriginCookie = ".origin";
 
-    public static void Handle(Config config, HttpRequest request, HttpResponse response)
+    public static void SetOriginRedirectUri(Config config, HttpResponse response, string origin)
     {
-        if (request.Cookies.ContainsKey(OriginCookie))
+        Globals.Logger.LogInformation("Origin header found - adding origin cookie");
+        response.Cookies.Append(OriginCookie, origin, new()
         {
-            Globals.Logger.LogInformation("Origin cookie already in place - returning");
-            return;
-        }
-
-        if (request.Headers.ContainsKey(OriginHeader))
-        {
-            Globals.Logger.LogInformation("Origin header found - adding origin cookie");
-            response.Cookies.Append(OriginCookie, request.Headers[OriginHeader], new()
-            {
-                Domain = config.CookieDomain,
-                Expires = DateTimeOffset.UtcNow.AddHours(1),
-                MaxAge = TimeSpan.FromHours(1),
-                SameSite = SameSiteMode.None,
-                HttpOnly = true,
-                Secure = true,
-                IsEssential = true
-            });
-        }
+            Domain = config.CookieDomain,
+            Expires = DateTimeOffset.UtcNow.AddHours(1),
+            MaxAge = TimeSpan.FromHours(1),
+            SameSite = SameSiteMode.None,
+            HttpOnly = true,
+            Secure = true,
+            IsEssential = true
+        });
     }
 
     public static void RedirectToOrigin(this HttpResponse response, HttpRequest request)
     {
         if (request.Cookies.ContainsKey(OriginCookie))
         {
-            response.Redirect(HttpUtility.UrlDecode(request.Cookies[OriginCookie]!));
+            var redirectUri = HttpUtility.UrlDecode(request.Cookies[OriginCookie]);
+            Globals.Logger.LogInformation($"Redirecting to {redirectUri} with queryString {request.QueryString}");
+            response.Redirect($"{redirectUri}{request.QueryString}");
         }
     }
 }
